* remotes/github/test-task: (35 commits)
  [task]fix vruntime when create task
  [task]using auto destroy task
  [task]add ps command
  [task]working on rb task
  [task]working on rb task, one task have a bug
  [task]working on rb task
  [task]working on rbtree task
  [task]remove task wait
  [task]remove runtime
  [task]working on task
  [task]add task_wait
  [task]add xfs_ctx for task
  [task]remove event_base_t in runtime
  [task]change stdin stdout and stderr to global
  [task]rename task stack size to stksz
  [task]move errno to task
  [task]working on task...
  [task]alloc scheduler in main function
  [smp]add smp.h and add smp_processor_id function
  [stddef]add likely and unlikely macro
  ...

# Conflicts:
#	src/framework/vm.c
#	src/include/xboot/event.h
#	src/include/xconfigs.h
#	src/init/main.c
#	src/kernel/command/cmd-ps.c
#	src/kernel/core/event.c
#	src/kernel/core/subsys.c
#	src/kernel/shell/shell.c
#	src/kernel/shell/system.c
#	src/kernel/xfs/xfs.c
#	src/lib/libc/environ/clearenv.c
#	src/lib/libc/environ/getenv.c
#	src/lib/libc/environ/putenv.c
#	src/lib/libc/environ/unsetenv.c

+++ src/kernel/core/subsys.c
@@ -39,9 +39,9 @@ static void subsys_init_romdisk(void)
 
     /* json = "{romdisk@0:{address:xxxxxxxx,size:yyyyyyyy}}" */
 	length = sprintf(json,
-		"{\"romdisk@0\":{\"address\":\"%lld\",\"size\":\"%lld\"}}",
-		(unsigned long long)(&__romdisk_start),
-		(unsigned long long)(&__romdisk_end - &__romdisk_start));
+		"{\"romdisk@0\":{\"address\":\"%ld\",\"size\":\"%ld\"}}",
+		(unsigned long)(&__romdisk_start),
+		(unsigned long)(&__romdisk_end - &__romdisk_start));
     /* 探测romdisk设备 */
 	probe_device(json, length, NULL);
 }
